# Improve Encoding.UTF8.GetString / GetChars performance for small inputs