@GrabYourPitchforks 2019-10-17
One consideration is that the performance of zero-length `byte[]` to `string` is regressed since the early-exit check was removed. Personally I'm fine with this, since zero-length `byte[]` inputs should be rare, and removing these checks makes the code cleaner. The `string` ctor is already checking for zero-length buffers and returning the `string.Empty` singleton, so this won't result in any extra allocations in this case.