@jkotas 2019-10-20
Marking blocked - depends on  mono/linker#792 being fixed.