@GrabYourPitchforks 2019-11-06
The mono linker issue is now fixed, currently just waiting for things to flow back to our build system.