@GrabYourPitchforks 2019-11-07
Spoke offline with Sven, and we expect Maestro bot to inject the new mono linker into coreclr tonight or early tomorrow morning. I can merge this PR immediately after that goes through.