@sbomer 2019-11-08
The subscription still isn't flowing the linker update even with https://github.com/dotnet/coreclr/pull/27736. @dotnet/dnceng would someone be able to look at why dependency flow from mono/linker to coreclr isn't working? `darc update-dependencies` works locally. `darc trigger-subscription` succeeds but still hasn't created a PR.