@riarenas 2019-11-08
I was investigating the missing dependency update PR and I manually reset the subscription's last Applied build and retriggered it to get a fresh set of logs. The retrigger opened https://github.com/dotnet/coreclr/pull/27773.

I'll see if I can find why maestro thought it had already applied the build without opening a PR first. I opened https://github.com/dotnet/arcade/issues/4314 for investigation.

@sbomer Let me know what you want to do with #27773 and #27771