@GrabYourPitchforks 2019-11-09
Confirmed that the correct codegen is now being produced by these methods. Thank you for your help!